some abstraction to cleanup code
