+++ public/app/features/playlist/PlaylistPage.tsx
@@ -1,28 +1,19 @@
 import React, { FC, useState } from 'react';
 import { connect, MapStateToProps } from 'react-redux';
-import { NavModel, SelectableValue, urlUtil } from '@grafana/data';
+import { NavModel } from '@grafana/data';
 import Page from 'app/core/components/Page/Page';
 import { StoreState } from 'app/types';
 import { GrafanaRouteComponentProps } from '../../core/navigation/types';
 import { getNavModel } from 'app/core/selectors/navModel';
 import { useAsync } from 'react-use';
-import { getBackendSrv, locationService } from '@grafana/runtime';
+import { getBackendSrv } from '@grafana/runtime';
 import { PlaylistDTO } from './types';
-import {
-  Button,
-  Card,
-  Checkbox,
-  ConfirmModal,
-  Field,
-  LinkButton,
-  Modal,
-  RadioButtonGroup,
-  VerticalGroup,
-} from '@grafana/ui';
+import { Button, Card, ConfirmModal, LinkButton } from '@grafana/ui';
 import { contextSrv } from 'app/core/services/context_srv';
 import PageActionBar from 'app/core/components/PageActionBar/PageActionBar';
 import EmptyListCTA from '../../core/components/EmptyListCTA/EmptyListCTA';
 import { deletePlaylist } from './api';
+import { StartModal } from './StartModal';
 
 interface ConnectedProps {
   navModel: NavModel;
@@ -40,6 +31,7 @@ export const PlaylistPage: FC<PlaylistPageProps> = ({ navModel }) => {
     return getBackendSrv().get('/api/playlists', { query: searchQuery }) as Promise<PlaylistDTO[]>;
   }, [forcePlaylistsFetch]);
   const hasPlaylists = playlists && playlists.length > 0;
+  const playlistItemToDelete = playlists?.find(({ id }) => id === playlistToDelete);
 
   let content = (
     <EmptyListCTA
@@ -84,8 +76,6 @@ export const PlaylistPage: FC<PlaylistPageProps> = ({ navModel }) => {
       </>
     );
   }
-  const playlistItemToDelete = playlists?.find(({ id }) => id === playlistToDelete);
-
   return (
     <Page navModel={navModel}>
       <Page.Contents isLoading={loading}>
@@ -125,52 +115,3 @@ const mapStateToProps: MapStateToProps<ConnectedProps, {}, StoreState> = (state:
 });
 
 export default connect(mapStateToProps)(PlaylistPage);
-
-export interface StartModalProps {
-  playlist: PlaylistDTO;
-  onDismiss: () => void;
-}
-
-export const StartModal: FC<StartModalProps> = ({ playlist, onDismiss }) => {
-  const [mode, setMode] = useState<any>(false);
-  const [autoFit, setAutofit] = useState(false);
-
-  const modes: Array<SelectableValue<any>> = [
-    { label: 'Normal', value: false },
-    { label: 'TV', value: 'tv' },
-    { label: 'Kiosk', value: true },
-  ];
-
-  const onStart = () => {
-    const params: any = {};
-    if (mode) {
-      params.kiosk = mode;
-    }
-    if (autoFit) {
-      params.autofitpanels = true;
-    }
-    locationService.push(urlUtil.renderUrl(`/playlists/play/${playlist.id}`, params));
-  };
-
-  return (
-    <Modal isOpen={true} icon="play" title="Start playlist" onDismiss={onDismiss}>
-      <VerticalGroup>
-        <Field label="Mode">
-          <RadioButtonGroup value={mode} options={modes} onChange={setMode} />
-        </Field>
-        <Checkbox
-          label="Autofit"
-          description="Panel heights will be adjusted to fit screen size"
-          name="autofix"
-          value={autoFit}
-          onChange={(e) => setAutofit(e.currentTarget.checked)}
-        />
-      </VerticalGroup>
-      <Modal.ButtonRow>
-        <Button variant="primary" onClick={onStart}>
-          Start {playlist.name}
-        </Button>
-      </Modal.ButtonRow>
-    </Modal>
-  );
-};
